requirements.txt
ae-ffmpeg==1.1.2
audioread==3.0.1
auto-editor==23.32.1
av==11.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.1.0
decorator==5.1.1
faster-whisper==1.0.1
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.3
packaging==24.0
Pillow==10.0.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==5.26.1
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tdqm==0.0.1
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
